requirements.txt
dnspython
pyyaml
apscheduler
python-redis    # optional, only if you use redis queue
requests
lxml
